feat(module) Add `Compile`, `Module.Instantiate*`, and `Module.Close`.
This patch adds a way to build a real `Module` by using the `Compile`
function. Then, a module can instantiated with `Module.Instantiate` or
`Module.InstantiateWithImports`.

The `NewInstanceWithImports` function has been updated to use a new
`newInstanceWithImports` private function that takes an “instance
builder” as a parameter, so that it's easier to share code `Instance`
and `Module` when it comes to create an `Instance` with its imports,
exports, and memory.