Add a test that shows the off-by-one error in the greet example
The error: the example doesn't allocate a byte for the NULL terminator,
so that terminator can get overwritten. In practice, playing around I
found this only caused a problem if I could get the input string to not
quite hit a block boundary in the linear memory, but I haven't worked
out exactly what is going on.