feat: Remove host function from the global store.
When the host function is dropped, remove it from the global store
too.

To avoid running out of value for the host function store index, we
use a “generational index”-ish approach. When a host function is
removed from the store, its index can be reused by another host
function.