feat: Avoid double-free of the same trap.
So first, `wasm_trap_t` are automatically dropped by Rust in our
case. But in case of, it is possible to have a double-free here. When
a host function is declared, the trampoline function transforms errors
to trap with `NewTrap`, so the destructor is attached to this new
pointer. And then, a Go error is built from this trap by calling
`newErrorFromTrap`, which calls `newTrap`, which attaches another
destructor to the same pointer. Usually we avoid such situation by
using the `ownedBy` field, but in this case, the trap is owned by
nobody. So… in case of, we prevent a double-free with this patch.