fix(function) Avoid calling `Function.Type` for every call.
When `Function.Call` is called, a new `FunctionType` is created with
`Function.Type`. The `FunctionType` is “owned by” `Function`, and
isn't freed until `Function` is freed.

When calling the same function again and again in a loop, all the
`FunctionType` will stack and increase the memory size.

By moving `FunctionType` outside the `lazyNative` function, a unique
`FunctionType` is created, thus saving a ton of calls and memory.