feat: Update to Wasmer 2.0.
This patch updates `wasmer-go` to the Wasmer runtime 2.0.

The support for Linux aarch64 has been disabled temporarily.
The support for Darwin aarch64 is new.

The `JIT` and `Native` engines have been renamed `Universal` and
`Dylib`. All related functions and methods are still present but are
marked as deprecated. Use the new `UseUniversalEngine`,
`UseDylibEngine`, `NewUniversalEngine`, and `NewDylibEngine`
functions/methods.